remove 'retry' loop in entrypoint
the retry loop was an attempt to get this container to always work but it doesn't solve the problem of the 'import hou' step failing occasionally.

instead, the whole process has been simplified. a few 'sleep's have been added (a bit of superstition in leu of an actual answer to the problem) and a small change to the 'HOUDINI_SCRIPT_LICENSE' ENV has been added.

additionally, updated the README for this issue